refactor: move RBAC manager from core module to rbac package
