+++ modules/core/presentation/controllers/dtos/role_dto.go
@@ -3,12 +3,15 @@ package dtos
 import (
 	"context"
 	"fmt"
-	"github.com/go-playground/validator/v10"
-	"github.com/google/uuid"
+
 	"github.com/iota-uz/iota-sdk/modules/core/domain/aggregates/role"
 	"github.com/iota-uz/iota-sdk/modules/core/domain/entities/permission"
 	"github.com/iota-uz/iota-sdk/pkg/composables"
 	"github.com/iota-uz/iota-sdk/pkg/constants"
+	"github.com/iota-uz/iota-sdk/pkg/rbac"
+
+	"github.com/go-playground/validator/v10"
+	"github.com/google/uuid"
 	"github.com/nicksnyder/go-i18n/v2/i18n"
 )
 
@@ -42,7 +45,7 @@ func (r *CreateRoleDTO) Ok(ctx context.Context) (map[string]string, bool) {
 	return errorMessages, len(errorMessages) == 0
 }
 
-func (r *CreateRoleDTO) ToEntity(rbac permission.RBAC) (role.Role, error) {
+func (r *CreateRoleDTO) ToEntity(rbac rbac.RBAC) (role.Role, error) {
 	perms := make([]*permission.Permission, 0, len(r.Permissions))
 	for permID := range r.Permissions {
 		permUUID, err := uuid.Parse(permID)
@@ -90,7 +93,7 @@ func (r *UpdateRoleDTO) Ok(ctx context.Context) (map[string]string, bool) {
 	return errorMessages, len(errorMessages) == 0
 }
 
-func (r *UpdateRoleDTO) ToEntity(roleEntity role.Role, rbac permission.RBAC) (role.Role, error) {
+func (r *UpdateRoleDTO) ToEntity(roleEntity role.Role, rbac rbac.RBAC) (role.Role, error) {
 	perms := make([]*permission.Permission, 0, len(r.Permissions))
 	for permID := range r.Permissions {
 		permUUID, err := uuid.Parse(permID)
